align definition with current OSPO growth
Open source program offices (OSPOs) can be established in a variety of sectors, regions, and organizational sizes, including private and public organizations. Therefore, it's important that we review and update the traditional OSPO definition to be more inclusive and up-to-date.

This PR aims to enhance the OSPO definition by incorporating a new section that focuses on OSPO segmentation based on the organization's goals and motivators. 

Signed-off-by: Ana Jimenez Santamaria <anitajs95@gmail.com>